Some commenting and changed cooldown on spotify
