+++ api/external/_getcredentials.ts
@@ -33,7 +33,7 @@ export async function getProviderCredentials(req: VercelRequest, res: VercelResp
 		)?.data;
 		return result;
 	} catch (e) {
-		console.error("PROVIDER (myapi): ", e, " msg: " + e?.response?.data?.message);
+		console.error("PROVIDER (myapi): ", e, " msg: " + e?.response?.data?.message, e?.data);
 		res.statusCode = e?.message?.includes("401") ? 401 : 404;
 		res.json({
 			ok: false,
