+++ src/components/NewWidgetModal.vue
@@ -12,7 +12,7 @@
 			<div
 				class="overflow-y-auto overflow-x-hidden fixed top-0 right-0 left-0 w-full md:inset-0 h-modal md:h-full flex justify-center items-center"
 				v-if="open"
-				v-show="dragging"
+				v-show="!dragging"
 				v-click-outside="toggle"
 			>
 				<div class="relative p-4 w-full max-w-2xl h-full md:h-auto z-100">
@@ -61,7 +61,7 @@ export default defineComponent({
 	data() {
 		return {
 			open: false,
-			dragging: true,
+			dragging: false,
 		};
 	},
 	computed: {
@@ -83,10 +83,10 @@ export default defineComponent({
 		},
 		generateKey,
 		start() {
-			this.dragging = false;
+			this.dragging = true;
 		},
 		end() {
-			this.dragging = true;
+			this.dragging = false;
 		},
 	},
 	setup() {
