+++ src/main.ts
@@ -1,15 +1,3 @@
-declare global {
-	interface Window {
-		env: {
-			VUE_APP_VERCEL_ENV: "production" | "preview" | "development";
-			VUE_APP_VERCEL_URL: string;
-			VUE_APP_VERCEL_GIT_COMMIT_SHA: string;
-			VUE_APP_VERCEL_GIT_COMMIT_MESSAGE: string;
-			BASEURL: string;
-		};
-	}
-}
-
 import { createApp } from "vue";
 import App from "./App.vue";
 import router from "./router";
@@ -41,7 +29,7 @@ import Toast from "vue-toastification";
 import "vue-toastification/dist/index.css";
 import { useUserStore } from "./store/user.store";
 
-//@ts-ignore
+//@ts-ignore This constant is inserted by webpack in vue.config.js
 window.env = ENV;
 window.env.VUE_APP_VERCEL_GIT_COMMIT_SHA = window.env.VUE_APP_VERCEL_GIT_COMMIT_SHA || "PLACEHOLDERAood4vTEZvU";
